Extracted exception advice for throwable to separate file
